Added Box class + tested reverse function calls
Made a box sub-class of entity and tested calling the resolveCollision function on it to see if it worked, it worked, and the box was still able to be pushed around by the player.
Problem is that the box being pushed was able to be pushed over the wall, that will need to be fixed. Error Picture was uploaded for this.