Fixed double resolving when using loop
I added a loop to stop using repeated lines when initialising objects, especially useful since my game will have a lot of these objects that I'll need to resolve, however, a slight logic error came up where I'm resolving the collision of the same 2 objects twice since I was using a nested for loop to loop over all the objects. To fix this, I am using a loop that uses an iterator and starting from the first element but starting the second one from the 2nd element in the list so that I don't resolve the collision a second time unnecessarily.