Adding walls and adding falling
Walls have been added using a tilemap with an array, but the main thing is that I needed to separate the collision checking from the walls and the normal objects. This is because the walls were constantly checking collisions with each other despite them never going to collide at all since they are stationary. I separated them using another loop especially for them and only resolving collisions vs them and the objects list.