Finished adding a walk through feature for objects
Commented all code that is not to do with collisions, next step will be to comment and further understand the collision code so that I can prepare to update it to fix the bugs. 1) Bugs with pushing the boxes, Bugs with player being rammed away as they run into a wall whilst jumping and on top of boxes and on the ground. After I remake the collision code try to remake this from scratch with little help.